Check for a queen attacking a queen left of the board
Since we are going from left to right and placing the queens then
there is no reason to check if the queen is attacking another queen
to the right.  Furthermore, check on the vertical is not necessary
since we will be placing a single queen on the files of the board.